aligning with bomb

+++ src/robot_control/include/robot_control/path_planning.h
@@ -1,7 +1,6 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <ros/ros.h>
-#include <unordered_set>
 #include <cmath>
 #include <fstream>
 #include <algorithm>
@@ -27,7 +26,7 @@ public:
 private:
 	void defineLocalPath(vector<int>* x_path, vector<int>* y_path);
 	inline vector<Node*>::iterator find(vector<Node*>* vec, int x, int y);
-	void writeMap(vector<unsigned char> map, float cell_size);
+	void writeMap(vector<unsigned char> map, double cell_size);
 	inline Node* initNode(int x, int y, int g, Node* parent);
 
 	double end_x, end_y;
